[BE] refactor S3 scripts and add new functionality for userbenchmarks
