+++ userbenchmark/utils.py
@@ -1,6 +1,6 @@
 import os
 import sys
-from datetime import datetime
+from datetime import datetime, timedelta
 import time
 import json
 from pathlib import Path
@@ -28,6 +28,10 @@ class add_path():
             sys.path.remove(self.path)
         except ValueError:
             pass
+
+
+with add_path(str(REPO_PATH)):
+    from utils.s3_utils import S3Client, USERBENCHMARK_S3_BUCKET, USERBENCHMARK_S3_OBJECT
 
 
 @dataclass
@@ -62,8 +66,45 @@ def dump_output(bm_name, output):
     with open(full_fname, 'w') as f:
         json.dump(output, f, indent=4)
 
+def get_date_from_metrics(metrics_file: str):
+    datetime_obj = datetime.strptime(metrics_file, "metrics-%Y%m%d%H%M%S")
+    return datetime.strftime(datetime_obj, "%Y-%m-%d")
+
+def get_ub_name(metrics_file_path: str):
+    with open(metrics_file_path, "r") as mf:
+        metrics = json.load(mf)
+    return metrics["name"]
+
 def get_output_dir(bm_name):
     current_dir = Path(os.path.dirname(os.path.abspath(__file__)))
     target_dir = current_dir.parent.joinpath(".userbenchmark", bm_name)
     target_dir.mkdir(exist_ok=True, parents=True)
     return target_dir
+
+
+def get_latest_n_jsons_from_s3(n: int, bm_name: str, platform_name: str, date: str):
+    """Retrieves the most recent n metrics json filenames from S3 the WEEK BEFORE the given date, exclusive of that date.
+       If fewer than n items are found, returns all found items without erroring, even if there were no items. """
+    s3 = S3Client(USERBENCHMARK_S3_BUCKET, USERBENCHMARK_S3_OBJECT)
+    directory = f'{bm_name}/{platform_name}'
+
+    if not s3.exists(None, directory):
+        return []
+
+    previous_json_files = []
+    start_date = datetime.strptime(date, '%Y-%m-%d')
+    current_date = start_date - timedelta(days=1)
+    while len(previous_json_files) < n and current_date >= start_date - timedelta(days=7):
+        current_date_str = current_date.strftime('%Y-%m-%d')
+        current_directory = f'{directory}/{current_date_str}'
+
+        if s3.exists(None, current_directory):
+            files = s3.list_directory(current_directory)
+            metric_jsons = [f for f in files if f.endswith('.json') and 'metrics' in f]
+            metric_jsons.sort(key=lambda x: datetime.strptime(x.split('/')[-1].split('-')[-1].split('.')[0], '%Y%m%d%H%M%S'), reverse=True)
+            previous_json_files.extend(metric_jsons[:n - len(previous_json_files)])
+
+        # Move on to the previous date.
+        current_date -= timedelta(days=1)
+
+    return previous_json_files
